+++ src/cltl/brain/commons/casefolding.py
