+++ src/cltl/brain/commons/discrete.py
