+++ src/cltl/brain/commons/triple_helpers.py
