+++ src/cltl/brain/reasoners/location_reasoner.py
@@ -2,7 +2,7 @@ import pathlib
 
 from cltl.brain.basic_brain import BasicBrain
 from cltl.brain.utils.helper_functions import read_query
-from cltl.combot.backend.utils.casefolding import casefold_text
+from cltl.brain.commons.casefolding import casefold_text
 
 
 class LocationReasoner(BasicBrain):
