+++ src/cltl/brain/utils/base_cases.py
@@ -1,7 +1,7 @@
 from datetime import date
 from random import getrandbits
 
-from cltl.combot.backend.api.discrete import UtteranceType
+from cltl.brain.commons.discrete import UtteranceType
 
 context_id, place_id = getrandbits(8), getrandbits(8)
 chat_1 = [
